+++ carveracontroller/addons/probing/ProbingPreviewAndRunPopup.kv
@@ -1,5 +1,5 @@
-<ProbingPopup>:
-    title:"Probing"
+<ProbingPreviewAndRunPopup>:
+    title:"Confirm Probing"
     size_hint: 0.7, 0.6
     pos_hint: {"center_x": 0.5, "center_y": 0.5}
     auto_dismiss: False
