--- carveracontroller/addons/probing/Tabs/MiscTab.kv
@@ -1,7 +1,0 @@
-<MiscTab@TabbedPanelItem>
-    text: tr._('Misc')
-    GridLayout:
-        cols: 3
-        rows: 10
-        spacing: dp(8)
-        padding: 0
