--- carveracontroller/addons/probing/Tabs/ProbeSettingsTab.kv
@@ -1,71 +1,0 @@
-
-# Note: i only did a tabbed panel here since I dont want to spend
-# the time to figure out making a title header :grimace:
-
-<ProbeSettingsTab@TabbedPanel>:
-    size_hint_max_x: root.size_hint_min_x
-    do_default_tab: False
-    tab_width: self.width
-    TabbedPanelItem:
-        text: tr._('Settings')
-        GridLayout:
-            cols:1
-            rows:10
-            spacing:dp(3)
-            padding:dp(1)
-
-            ProbeSettingLabel:
-                label_text: 'Tip Dia:'
-                hint_text: 'tip diameter'
-
-            ProbeSettingLabel:
-                label_text: 'Slow Feed Rate:'
-                hint_text: 'Slow Feed Rate'
-
-            ProbeSettingLabel:
-                label_text: 'Fast Feed:'
-                hint_text: 'Fast feed rate'
-
-            ProbeSettingLabel:
-                label_text: 'Max Distance:'
-                hint_text: 'Max Probing distance'
-                visible:False
-
-#Button:
-#    id: btn_ok
-#    text: tr._('Start')
-#    on_release:
-#        root.coord_popup.set_config(ProbingConstants.config_section, ProbingConstants.probe_switch_type, root.get_probe_switch_type())
-#        root.coord_popup.set_config(ProbingConstants.config_section, 'x', txt_x.text)
-#        root.coord_popup.set_config(ProbingConstants.config_section, 'y', txt_y.text)
-#        root.coord_popup.set_config(ProbingConstants.config_section, 'z', txt_z.text)
-#        root.coord_popup.set_config(ProbingConstants.config_section, 'a', txt_a.text)
-#
-#        app.root.start_probing(txt_x.text, txt_y.text, txt_z.text, txt_a.text, root.get_probe_switch_type())
-#        #root.dismiss()
-
-<ProbeSettingLabel@GridLayout>:
-    cols:2
-    rows:1
-    spacing: dp(3)
-    padding: dp(3)
-    size_hint_max_x: dp(300)
-    size_hint_max_y: dp(33)
-    label_text:''
-    hint_text:''
-
-    Label:
-        text: root.label_text
-        size_hint_min_x: dp(75)
-        halign: 'right'
-        valign: 'middle'
-        text_size: self.size
-
-    TextInput:
-        hint_text: root.hint_text
-        multiline: False
-        halign:'left'
-        input_type: 'number'
-        input_filter: 'float'
-
-
