--- carveracontroller/addons/probing/Tabs/ProbingTabs.kv
@@ -1,35 +1,0 @@
-
-#:include addons/probing/tabs/SingleAxisTab.kv
-#:include addons/probing/tabs/CornersTab.kv
-#:include addons/probing/tabs/BoreBossTab.kv
-#:include addons/probing/tabs/MiscTab.kv
-
-<ProbingTabs@TabbedPanel>
-    size_hint_min_x: root.size_hint_min_x
-    do_default_tab: False
-    tab_width: dp(111)
-
-    SingleAxisTab:
-    CornersTab:
-    BoreBossTab:
-    MiscTab:
-
-
-<ProbeButton@Button>:
-    background_color: 0,0,0,0  # the last zero is the critical on, make invisible
-    size: dp(100), dp(100)
-    size_hint: None, None
-    image: ''
-    canvas.before:
-        Color:
-            rgba: (88/255,88/255,88/255,1) if self.state=='normal' else (50/255,164/255,206/255,1)  # visual feedback of press
-        RoundedRectangle:
-            pos: self.pos
-            size: self.size
-            radius: [dp(3),]
-    Image:
-        source: root.image
-        size: root.size
-        size_hint: None, None
-        center_x: self.parent.center_x
-        center_y: self.parent.center_y
